Address review comments
Reverted accidental breaking change.
Move WalkContext to its own file
Removed limit property and made it a local variable.
