+++ src/language/walker/index.ts
@@ -21,4 +21,5 @@ export * from "./ruleWalker";
 export * from "./scopeAwareRuleWalker";
 export * from "./skippableTokenAwareRuleWalker";
 export * from "./syntaxWalker";
+export * from "./walkContext";
 export * from "./walker";
